4/30/22 1:54

+++ Coding Tracker 1/Program.cs
@@ -24,7 +24,7 @@ namespace Coding_Tracker_1
                     Date TEXT,
                     StartTime TEXT,
                     EndTime TEXT,
-                    Duration INT)";
+                    Duration REAL)";
 
                 tableCommand.ExecuteNonQuery();
